requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning Libraries
scikit-learn>=1.3.0
tensorflow-cpu>=2.13.0
joblib>=1.3.0

# Natural Language Processing
nltk>=3.8.0
wordcloud>=1.9.0

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
